pyproject.toml
[tool.poetry]
name = "karmagrambot"
version = "0.1.0"
description = "Telegram bot that logs messages (not the contents, for the sake of privacy) and votes."
authors = ["Cauê Baasch de Souza <cauebs@pm.me>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^11.1"
dataset = "^1.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
jedi = "^0.13.2"
autopep8 = "^1.4"

[tool.poetry.scripts]
karmagrambot = "karmagrambot:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
